# Faster hmac sha1 token generation