--- a/common/totp.js
+++ b/common/totp.js
@@ -1,45 +1,5 @@
 import {hmac_sha1} from "./hmac-sha1.js";
 
-let base32chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZ234567";
-
-// TODO move this to `tokens.js` instead of spending 100ms computing it every time a token is generated
-// Convert a base32-encoded string to a Uint8Array
-var base32touint8array = function(base32) {
-  base32 = base32.replace(/ +/g, ""); // Strip whitespace
-
-  let array = new Uint8Array(Math.floor(base32.length * 5 / 8));
-  let arrayIndex = 0;
-  let bits = 0;
-  let value = 0;
-
-  for (let i=0; i<base32.length; i++) {
-    let char = base32.charAt(i).toUpperCase();
-    if (char === '0') {
-        char = 'O';
-    } else if (char === '8') {
-        char = 'B';
-    } else if (char === '1') {
-        char = 'L';
-    }
-    let val = base32chars.indexOf(char);
-    if ( val >= 0 && val < 32 ) {
-      value = (value << 5) | val;
-      bits += 5;
-
-      // Transfer a byte into the Uint8Array if there is enough data
-      if (bits >= 8) {
-        array[arrayIndex++] = (value >>> (bits - 8)) & 0xFF;
-        value = value & 0xFF;
-        bits -= 8;
-      }
-    } else {
-      throw Error("Character out of range: " + char);
-    }
-  }
-
-  return array;
-};
-
 // Convert a TOTP counter to a Uint8Array
 var countertouint8array = function(counter) {
   let array = new Uint8Array(8);
@@ -56,9 +16,8 @@ var countertouint8array = function(counter) {
 
 export function TOTP() {
   this.getOTP = function(secret, epoch) {
-    let secretarray = base32touint8array(secret);
     let time = countertouint8array(Math.floor(epoch / 30));
-    let hmac = hmac_sha1(secretarray, time);
+    let hmac = hmac_sha1(secret, time);
 
     let offset = hmac[hmac.length - 1] & 0x0f;
     let otp = ((hmac[offset] << 24 | hmac[offset+1] << 16 | hmac[offset+2] << 8 | hmac[offset+3]) & 0x7fffffff) + "";