… common/sha.js