Switch from the [jsSHA](https://github.com/Caligatio/jsSHA) library to [js-sha1](https://github.com/emn178/js-sha1), which is roughly twice as fast on my watch. `js-sha1` doesn't have HMAC support built in, so that's now a custom implementation in `common/hmac-sha1.js`. Using `Uint8Array` everywhere instead of encoding and decoding hex/base32 strings gives a significant performance boost while computing tokens.

To avoid decoding the tokens from base32 each time the app launches, this PR changes the stored CBOR file from storing tokens as base32-encoded strings to storing the decoded byte arrays. Existing stored tokens will be converted when the app is first launched, and new tokens will be converted when they are added in settings. It would be possible to re-encode the byte arrays back to base32 if you want to switch out the libraries for OTP/HMAC/SHA1/etc. in the future.

For me, when I have 3 codes the current app takes around 4s to initially launch, and 2.5s to regenerate codes (tested by turning the screen off and on). This branch brings that to around 2.5s for the initial launch and less than 1s to regenerate codes.