@Lixxia 2019-06-08
Hiya, sorry been on vacation past few weeks. This looks great! I'll take a look this weekend and run some tests. Maybe you could give me a quick overview of which situations you tested?