@nhinds 2019-06-08
I tested starting with a couple of existing codes from AWS (which seems to have longer than usual codes) and one shorter code, and those got picked up and turned into the array buffer format. I also removed them and added the same codes back from the settings page to test that saving new codes is okay.

I've been using this version of the app myself for a couple of weeks and the codes have been accepted, so at least for AWS the codes are still being generated correctly.