@Lixxia 2019-06-08
Overall looks good, very solid approach and definitely reduces speeds on load/token regeneration. I've added a few comments in-line regarding a reintroduced bug and some code style preferences. Let me know if you have any questions!