@nhinds 2019-06-14
Re-tested by removing all codes, reverting to `master`, loading in a bunch of codes (including a normal-sized code, AWS codes, and the 26-character test code from `app/index.js`), and switching back over to this branch. All the codes got migrated over to uint8array format.

Also tested removing and re-adding the same codes as above, which seemed to work fine.