@Lixxia 2019-08-04
Hey, not sure if you're still looking at this or not. I did a fresh clone and install of your branch and seems there's still a few issues. There's a leftover reference to the `nodeWrap` (`p`) function in js-sha1.js that causes a `ReferenceError`. A quick attempt at trying to remove it and re-minify causes some other token handling errors down the road. Mind taking another look?