@nhinds 2019-08-05
It has been working fine for me for months after the latest push, not sure what would be causing reference errors. The nodeWrap/p function should never be called in Fitbit because it's guarded by a check for node.js, which should not be true for Fitbit's js interpreter... Is something broken or are there logs somewhere that show the errors? 

I'll take a closer look later this week when I have time to debug.