@nhinds 2019-08-07
I found what you were referring to. It worked fine on my watch and phone, but I got this error when running under the simulator:
```
[8:23:36 PM] Companion: ReferenceError: p is not defined
  c at common/js-sha1.js:16,925
  ? at common/js-sha1.js:16,5266
  ? at common/js-sha1.js:16,2
```

Looks like the simulator's companion is running under node.js, or at least something that looks like node.js to the js-sha1 library. I've added the settings which disable node.js/commonjs support in js-sha1 and it seems to work fine in the simulator now. Are you able to check whether it is working for you as well?